JsonStringLocalizerBase created to manage them all :) #18
