+++ Askmethat.Aspnet.JsonLocalizer/Askmethat.Aspnet.JsonLocalizer.csproj
@@ -13,7 +13,7 @@
     <Copyright>Copyright 2018 (c) Askmethat Corporation.</Copyright>
     <PackageTags>Json Localizer Globalization netcore netstandard</PackageTags>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.1.3-preview1</Version>
+    <Version>1.1.4</Version>
     <RepositoryUrl>https://github.com/AlexTeixeira/Askmethat-Aspnet-JsonLocalizer</RepositoryUrl>
     <RepositoryType>Git</RepositoryType>
   </PropertyGroup>
