+++ test/Askmethat.Aspnet.JsonLocalizer.Test/Askmethat.Aspnet.JsonLocalizer.Test.csproj
@@ -5,7 +5,7 @@
 
     <IsPackable>false</IsPackable>
 
-    <Version>1.1.3-preview1</Version>
+    <Version>1.1.4</Version>
   </PropertyGroup>
 
   <ItemGroup>
