+++ test/Askmethat.Aspnet.JsonLocalizer.TestSample/Askmethat.Aspnet.JsonLocalizer.TestSample.csproj
@@ -1,42 +1,34 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp1.1;netcoreapp2.0;netcoreapp2.1;</TargetFrameworks>
-    <Version>1.1.3-preview1</Version>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="json\localization.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-
-  <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp1.1'">
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.3" />
-    <PackageReference Include="Microsoft.Composition" Version="1.0.27" ExcludeAssets="All" />
-    <PackageReference Include="System.Composition" Version="1.0.31" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp2.0'">
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.0" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp2.1'">
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.0 " />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.0 " />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Askmethat.Aspnet.JsonLocalizer\Askmethat.Aspnet.JsonLocalizer.csproj" />
-  </ItemGroup>
-
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<Project Sdk="Microsoft.NET.Sdk.Web">
+    <PropertyGroup>
+        <TargetFrameworks>netcoreapp1.1;netcoreapp2.0;netcoreapp2.1;</TargetFrameworks>
+        <Version>1.1.4</Version>
+    </PropertyGroup>
+    <ItemGroup>
+        <Folder Include="wwwroot\" />
+    </ItemGroup>
+    <ItemGroup>
+        <None Include="json\localization.json">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+    <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp1.1'">
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="1.1.2" />
+        <PackageReference Include="Microsoft.AspNetCore" Version="1.1.3" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.3" />
+        <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.3" />
+        <PackageReference Include="Microsoft.Composition" Version="1.0.27" ExcludeAssets="All" />
+        <PackageReference Include="System.Composition" Version="1.0.31" />
+    </ItemGroup>
+    <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp2.0'">
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.0" />
+    </ItemGroup>
+    <ItemGroup Condition="'$(TargetFramework)'=='netcoreapp2.1'">
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.0 " />
+        <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.0 " />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\..\Askmethat.Aspnet.JsonLocalizer\Askmethat.Aspnet.JsonLocalizer.csproj" />
+    </ItemGroup>
 </Project>
